+++ Source/Ursine3D/src/Core/Graphics/Animation/Builder/AnimationState.cpp
@@ -36,17 +36,17 @@ namespace ursine
 		m_name = name;
 	}
 	
-	double AnimationState::GetTimePosition(void) const
+	float AnimationState::GetTimePosition(void) const
 	{
 		return m_timePos;
 	}
 
-	void AnimationState::SetTimePosition(const double position)
+	void AnimationState::SetTimePosition(const float position)
 	{
 		m_timePos = position;
 	}
 
-	void AnimationState::IncrementTimePosition(const double dt)
+	void AnimationState::IncrementTimePosition(const float dt)
 	{
 		m_timePos += dt;
 	}
@@ -106,13 +106,37 @@ namespace ursine
 		m_animation = targetAnimation;
 	}
 
-	const double &AnimationState::GetTransPosition(void) const
+	const float &AnimationState::GetTransPosition(void) const
 	{
 		return m_transPos;
 	}
 
-	void AnimationState::SetTransPosition(const double& tPos)
+	void AnimationState::SetTransPosition(const float& tPos)
 	{
 		m_transPos = tPos;
 	}
+
+	//void AnimationState::GetFrameByTime(std::vector<AnimationKeyframe> &f1, std::vector<AnimationKeyframe> &f2, float time) const
+	//{
+	//	if (m_animation)
+	//		return;
+	//
+	//	// determine the 2 current keyframes to use
+	//	// we assume that all frames exist, and that they were baked across all total keyframes
+	//	unsigned frameCount = m_animation->GetRigKeyFrameCount();
+	//	for (unsigned x = 0; x < frameCount - 1; ++x)
+	//	{
+	//		// get the two current keyframes
+	//		const std::vector<AnimationKeyframe> &frame1 = m_animation->GetKeyframes(x);
+	//		const std::vector<AnimationKeyframe> &frame2 = m_animation->GetKeyframes(x + 1);
+	//
+	//		// check if the current keyframe set holds the time value between them
+	//		if (frame1[0].length <= time && time < frame2[0].length)
+	//		{
+	//			f1 = frame1;
+	//			f2 = frame2;
+	//			break;
+	//		}
+	//	}
+	//}
 }
